+++ R/DD.dict.T.R
@@ -1,0 +1,7 @@
+#' @title DD.dict.T
+#' @name DD.dict.T
+#' @description Data dictionary embedded in ExampleT.
+#' @usage data(ExampleT)
+#' @docType data
+#' @seealso \code{\link{ExampleT}}
+NULL