+++ R/DS.data.T.R
@@ -1,0 +1,7 @@
+#' @title DS.data.T
+#' @name DS.data.T
+#' @description Data set embedded in ExampleT.
+#' @usage data(ExampleT)
+#' @docType data
+#' @seealso \code{\link{ExampleT}}
+NULL