+++ R/ExampleT.R
@@ -1,0 +1,12 @@
+#' @title ExampleT
+#' @name ExampleT
+#' @description Example data set and data dictionary with intentional errors.
+#' @usage data(ExampleT)
+#' @docType data
+#' @format \code{R data file} that contains two objects:
+#' \describe{
+#' \item{DD.dict.T}{Data dictionary}
+#' \item{DS.data.T}{Data set}
+#' }
+#' @keywords datasets
+NULL